Updating directory structure
