+++ public/index.html
@@ -28,6 +28,9 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <!--Start code-->
+    <div id="root">
+    </div>
+    <!--End code-->
   </body>
 </html>
