--- src/App.css
@@ -1,7 +1,0 @@
-.container {
-  text-align: center;
-  height: 100vh;
-  width: 90vw;
-  margin: 0 auto;
-  color: rgb(00, 224, 100);
-}
