+++ src/components/Header/Header.jsx
@@ -1,4 +1,5 @@
 import React, {Component} from 'react';
+import './Header.pip.css';
 
 
 class Header extends Component {
@@ -9,7 +10,11 @@ class Header extends Component {
     render() {
         return (
             <div>
-                    
+                <nav>
+                    <span>INTRO</span>
+                    <span>PROJECTS</span>
+                    <span>MORE</span>
+                </nav>   
             </div>
         )
     }
