+++ src/components/Header/Header.pip.css
@@ -1,0 +1,19 @@
+nav {
+    height: 10vh;
+    width: 100vw;
+    position: absolute;
+    top: 0;
+    left: 0;
+    padding-left: 10%;
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+    border-bottom: 1px solid #8df776;
+}
+
+nav span {
+    line-height: 31px;
+    font-size: 1.25rem;
+    letter-spacing: 0.3px;
+    margin-left: 14px;
+    font-weight: 300;
+}
